Updated ContactEmail Test class

+++ src/main/java/com/leadlet/service/impl/ContactEmailServiceImpl.java
@@ -1,5 +1,6 @@
 package com.leadlet.service.impl;
 
+import com.leadlet.security.SecurityUtils;
 import com.leadlet.service.ContactEmailService;
 import com.leadlet.domain.ContactEmail;
 import com.leadlet.repository.ContactEmailRepository;
@@ -12,13 +13,15 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import javax.persistence.EntityNotFoundException;
+
 
 /**
  * Service Implementation for managing ContactEmail.
  */
 @Service
 @Transactional
-public class ContactEmailServiceImpl implements ContactEmailService{
+public class ContactEmailServiceImpl implements ContactEmailService {
 
     private final Logger log = LoggerFactory.getLogger(ContactEmailServiceImpl.class);
 
@@ -41,46 +44,74 @@ public class ContactEmailServiceImpl implements ContactEmailService{
     public ContactEmailDTO save(ContactEmailDTO contactEmailDTO) {
         log.debug("Request to save ContactEmail : {}", contactEmailDTO);
         ContactEmail contactEmail = contactEmailMapper.toEntity(contactEmailDTO);
+        contactEmail.setAppAccount(SecurityUtils.getCurrentUserAppAccount());
         contactEmail = contactEmailRepository.save(contactEmail);
         return contactEmailMapper.toDto(contactEmail);
     }
 
     /**
-     *  Get all the contactEmails.
+     * Update a contactEmail.
      *
-     *  @param pageable the pagination information
-     *  @return the list of entities
+     * @param contactEmailDTO the entity to update
+     * @return the persisted entity
+     */
+    @Override
+    public ContactEmailDTO update(ContactEmailDTO contactEmailDTO) {
+        log.debug("Request to update ContactEmail : {}", contactEmailDTO);
+
+        ContactEmail contactEmail = contactEmailMapper.toEntity(contactEmailDTO);
+        ContactEmail contactEmailFromDb = contactEmailRepository.findOneByIdAndAppAccount(contactEmail.getId(), SecurityUtils.getCurrentUserAppAccount());
+
+        if (contactEmailFromDb != null) {
+            contactEmail.setAppAccount(SecurityUtils.getCurrentUserAppAccount());
+            contactEmail = contactEmailRepository.save(contactEmail);
+            return contactEmailMapper.toDto(contactEmail);
+        } else {
+            throw new EntityNotFoundException();
+        }
+    }
+
+    /**
+     * Get all the contactEmails.
+     *
+     * @param pageable the pagination information
+     * @return the list of entities
      */
     @Override
     @Transactional(readOnly = true)
     public Page<ContactEmailDTO> findAll(Pageable pageable) {
         log.debug("Request to get all ContactEmails");
-        return contactEmailRepository.findAll(pageable)
+        return contactEmailRepository.findByAppAccount(SecurityUtils.getCurrentUserAppAccount(), pageable)
             .map(contactEmailMapper::toDto);
     }
 
     /**
-     *  Get one contactEmail by id.
+     * Get one contactEmail by id.
      *
-     *  @param id the id of the entity
-     *  @return the entity
+     * @param id the id of the entity
+     * @return the entity
      */
     @Override
     @Transactional(readOnly = true)
     public ContactEmailDTO findOne(Long id) {
         log.debug("Request to get ContactEmail : {}", id);
-        ContactEmail contactEmail = contactEmailRepository.findOne(id);
+        ContactEmail contactEmail = contactEmailRepository.findOneByIdAndAppAccount(id, SecurityUtils.getCurrentUserAppAccount());
         return contactEmailMapper.toDto(contactEmail);
     }
 
     /**
-     *  Delete the  contactEmail by id.
+     * Delete the  contactEmail by id.
      *
-     *  @param id the id of the entity
+     * @param id the id of the entity
      */
     @Override
     public void delete(Long id) {
         log.debug("Request to delete ContactEmail : {}", id);
-        contactEmailRepository.delete(id);
+        ContactEmail contactEmailFromDb = contactEmailRepository.findOneByIdAndAppAccount(id, SecurityUtils.getCurrentUserAppAccount());
+        if (contactEmailFromDb != null) {
+            contactEmailRepository.delete(id);
+        } else {
+            throw new EntityNotFoundException();
+        }
     }
 }
